feat(core): Adding the ability to load an instance of the game engine from a jsonfile or dict. (#54)
## Describe your changes

We want to ensure we can serialize and deserialize the game engine into
a json format to allow us to start savings game.

## Issue ticket number and link

Closes #27 

## Checklist before requesting a review
- [ ] I have performed a self-review of my code
- [ ] I have defined what tests cover my change and or added new test to
cover change.
- [ ] I have added comments to the change and updated any relevant docs.
- [ ] I have pulled the latest in the main branch and merged them ahead
of time with no conflicts.

---------
